Adding updating changes based on updating modifiers
The modifiers will now update and change the skills according to the difference. The basic concept of this app is now complete. I want to change the new character page to be more intuitive and guide newcomers through the creation process. Also more features for the character stats screen like calculators and notes, spell page, background, feats, etc.