A bit of a revmap
Got rid of a lot of useless functions that didn't need to exist. Got rid of a global dependency that was unnecessary. Seriously globals are just a bad idea to use Michael! Also added new code to set up for a potential spell page in the future. All cantrip spells are already added for parsing into a spell database! Woot. Needs a actual spell page for the users and adding them to tie into personal characters. Lots of work to do for a spell page.... Overall this is a lot cleaner code.